Updated readme.md to correct finalAPY calculation
The "totalAPY" formula provided in the readme does not reflect the formula used by getFarmWithTradingFeesApy.ts, as it indicates that tradingAPR is compounded rather than vaultAPR. This update replaces vaultAPR with vaultAPY and removes references to tradingAPR being compounded. It also adds a short section directing readers to the new API docs page, which contains more details on additional endpoints and is actively being maintained.